% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergm.R
\name{AIC.sergm}
\alias{AIC.sergm}
\title{AIC for sergm Objects}
\usage{
\method{AIC}{sergm}(object, ..., k)
}
\arguments{
\item{object}{A fitted tsergm model (object of class \code{sergm}).}

\item{...}{optionally more fitted model objects (not used at the moment).}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is the classical AIC.#' @return numeric value of AIC.}
}
\description{
Prints the AIC of an estimated sergm model.
}
