% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsergm.R
\name{AIC.tsergm}
\alias{AIC.tsergm}
\title{AIC for tsergm Objects}
\usage{
\method{AIC}{tsergm}(object, ..., k = 2)
}
\arguments{
\item{object}{A fitted tsergm model (object of class \code{tsergm}).}

\item{...}{optionally more fitted model objects (not used at the moment).}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is the classical AIC.}
}
\value{
numeric value of AIC.
}
\description{
Prints the AIC of an estimated sergm model.
}
