% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R, R/print.R
\name{control.sergm}
\alias{control.sergm}
\alias{print.control.sergm}
\title{Control the specifications of the estimation algorithm of a Signed Exponential Random Graph Models}
\usage{
control.sergm(
  method_est = "Stepping",
  method_var = NULL,
  eval_likelihood = TRUE,
  mple_init = TRUE,
  max_it = 10,
  tol = 0.003,
  RM_c = 0.5,
  Stepping_number_grids = 20,
  n_bridges = 16,
  sampler_est = NULL,
  sampler_var = NULL,
  sampler_path_sampling = NULL,
  cluster = NULL
)

\method{print}{control.sergm}(x, ...)
}
\arguments{
\item{method_est}{There are currently four estimation algorithms implemented. \code{MLE} is the standard MCMCLE estimation routine 
with a log-normal approximation as proposed by Hummel et al. for binary networks. 
\code{Stepping} implements the stepping algorithm also introduced by by Hummel et al., 
while \code{RM} evokes the Robins Monroe algorithm for signed networks. 
Finally, the pseudo likelihood estimates based on a multinomial approximation 
of the intractable likelihood can be obtain through the option \code{MPLE}. 
For the pseudo likelihood the standard errors are approximated by a parametric 
bootstrap procedure introduced by Cranmer and Desmarais for static binary ERGMs.}

\item{method_var}{Method to be used to assess the uncertainty of the estimates. There are three possible options 
to be chosen from. The first option is "Fisher" and the standard choice, being based on approximating 
the Fisher information via the sampler given in the argument \code{sampler_var}. Second, 
"BS" translates to the bootstrap approximation detailed in Desmarais and Cranmer (2012). 
Finally, the variance of the MPLE can be used when the parameter equals "MPLE", although here some caution is in place as they are 
often too slim and their behavior is not studied in theoretical contexts. If no parameter is specified the default 
paramter is the "BS" option for estimations carried out via "MPLE" and "Fisher" for all MCMC-based estimation routines.}

\item{eval_likelihood}{Boolean indicator whether the logarithmic likelihood should be evaluated or not.}

\item{mple_init}{Boolean indicator whether the MPLE estimator or a vector of 
zeroes should be used as the initial estimator.}

\item{max_it}{Maximal number of iterations of the iterative procedure.}

\item{tol}{Define the tolerance of relative change in parameter vector between consecutive 
iterations to indicate convergence of the estimates.}

\item{RM_c}{If method = "RM" the gain sequence a_i in the i-th iteration is given by c^(-i).}

\item{Stepping_number_grids}{If method = "Stepping" the number of grids between 0 and 1 for which it will be tested 
whether the convex combination between the observed sufficient statistics and the mean simulated statistics lies 
in the convex hull of the samples of sufficient statistics.}

\item{n_bridges}{Number of bridges used for evaluating the log likelihood via path sampling.}

\item{sampler_est}{An \code{\link{sampler.sergm}} to specify the sampler used for the estimation of the parameters 
This parameter is not of importance if the chosen method is the MPLE.}

\item{sampler_var}{An \code{\link{sampler.sergm}} to specify the sampler used for assessing the uncertainty
of the parameters of the estimated \code{sergm}.}

\item{sampler_path_sampling}{An \code{\link{sampler.sergm}} to specify the sampler used for evaluating the log likelihood at the estimated parameter value. 
If the sampler is set to NULL the configurations of \code{sampler_var} are copied.}

\item{cluster}{A PSOCK or Fork cluster generated with \code{\link{makeCluster}} from the package \code{\link{parallel}}.}

\item{x}{See [print()].}

\item{...}{Other unused parameters
Automatically called when an object of class \code{\link{control.sergm}} is printed.}
}
\value{
Object of class \code{sampler.sergm}.
}
\description{
\code{control.sergm}  is used to specify how the estimation of a Signed 
Exponential Random Graph Models should be carried out.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: 
}}

