% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{count_statistics}
\alias{count_statistics}
\title{Calculate the global statistics given in the formula in the signed network on the lhs}
\usage{
count_statistics(formula)
}
\arguments{
\item{formula}{formula An \R \code{\link{formula}} object, of the form
\code{y ~ <model terms>}, where \code{y} is a numeric matrix including 
the signed entries 0, 1, and -1. Which terms are implemented is detailed in \code{\link{sergm.terms}}}
}
\description{
Calculate the global statistics given in the formula in the signed network on the lhs
}
