% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{find_min_in_ch}
\alias{find_min_in_ch}
\title{Function to find the gamma in the stepping algorithm as high as possible while still being in the convex hull of the sampled statistics}
\usage{
find_min_in_ch(gammas, obs_statistics, M, M_bar)
}
\arguments{
\item{gammas}{Vector of gamma values to be tried out}

\item{obs_statistics}{Matrix of simulated statistics (columns the separate sufficient statistics and rows the separate simulations)}

\item{M}{Observed simulated statistics}

\item{M_bar}{Average simulated statistics over all simulations}
}
\description{
Function to find the gamma in the stepping algorithm as high as possible while still being in the convex hull of the sampled statistics
}
