% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{map_to_mat}
\alias{map_to_mat}
\title{Convert a map to a matrix}
\usage{
map_to_mat(edges_pos, edges_neg, n_actors)
}
\arguments{
\item{edges_pos}{List of length n_actors that gives the positive edges for each actor}

\item{edges_neg}{List of length n_actors that gives the negative edges for each actor}

\item{n_actors}{Numeric value indicating the number of actors in the network}
}
\value{
igraph network.
}
\description{
\code{map_to_mat}  is used to covert a map (which is a c++ data structure) equal to a list to an adjacency matrix representing a signed network.
}
