% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{mat_to_sign_igraph}
\alias{mat_to_sign_igraph}
\title{Convert a matrix to an igraph object}
\usage{
mat_to_sign_igraph(mat)
}
\arguments{
\item{mat}{numeric matrix}
}
\value{
igraph network.
}
\description{
\code{mat_to_sign_igraph}  is used to covert an adjacency matrix representing a signed network to an igraph object.
}
