% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergm.R
\name{model_assessment}
\alias{model_assessment}
\title{Model assessment for sergm Objects}
\usage{
model_assessment(object, sampler = ergm.sign::sampler.sergm())
}
\arguments{
\item{object}{A fitted sergm model (object of class \code{\link{sergm}}).}

\item{sampler}{A fitted sergm model (object of class \code{\link{sampler.sergm}}).}
}
\value{
numeric value of AIC.
}
\description{
Samples from the converged object and then calculate the distribution of 
edgewise shared friends and enemies as well as positive and negative degree distributions.
}
