% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{parallel_sample}
\alias{parallel_sample}
\title{Parallel Sampling}
\usage{
parallel_sample(
  tmp_coef,
  terms,
  n_actors,
  networks,
  data_lists,
  type_list,
  mh,
  n_proposals,
  n_proposals_burn_in,
  seed,
  number_networks,
  cluster,
  start_with_empty_net
)
}
\arguments{
\item{tmp_coef}{numeric vector of coefficients to be sampled from}

\item{terms}{set of terms}

\item{n_actors}{Numeric value of the number of actors}

\item{networks}{List of signed networks (each represeting a snaptshot in time)}

\item{data_lists}{List of data matrices for each term}

\item{type_list}{List of type arguments for each term}

\item{mh}{boolean value indicating whether Metropolis-Hastings or Gibbs sampling should be used}

\item{n_proposals}{Number of iterations between sampled networks (number of tries)}

\item{n_proposals_burn_in}{Number of burn in iterations}

\item{seed}{Numeric seed to make the stochastic sampling reproducible}

\item{number_networks}{Numeric value of the number of networks to be sampled.}

\item{cluster}{A PSOCK or Fork cluster generated with \code{\link{makeCluster}} from the package \code{\link{parallel}}.}

\item{start_with_empty_net}{boolean value indicating whether we should start with the empty or observed network}
}
\value{
Object of class \code{sergm}.
}
\description{
\code{\link{parallel_sample}}  is used to sample temporal networks in parallel. This is an internal function and not to be used directly.
}
