% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R, R/print.R
\name{sampler.sergm}
\alias{sampler.sergm}
\alias{print.sampler.sergm}
\title{Sampler for Signed Exponential Random Graph Models}
\usage{
sampler.sergm(
  n_proposals_burn_in = 10000 * 20,
  n_proposals = 10000,
  seed = NA,
  number_networks = 100,
  init_empty = T,
  mh = T
)

\method{print}{sampler.sergm}(x, ...)
}
\arguments{
\item{n_proposals_burn_in}{Number of iterations of the sampler to be carried out 
as the burn-in period. This number should be picked to be high enough to reach the 
stationary distribution of the SERGM from which one wants to sample from.}

\item{n_proposals}{Number of iterations of the sampler to be carried out 
after the burn-in period for each network (often called MCMC interval).}

\item{seed}{Seed for the samples to make the code reproducible.}

\item{number_networks}{Number of networks to sample.}

\item{init_empty}{Boolean value indicating whether the initial network should be the observed network or an empty network.}

\item{mh}{Boolean value indicating whether an Metropolis Hastings or Gibbs Sampler should be used. 
If the value is TRUE the Metropolis Hastings algorithm is used, otherwise the Gibbs Sampler.}

\item{x}{See [print()].}

\item{...}{Other unused parameters
Automatically called when an object of class \code{\link{sampler.sergm}} is printed.}
}
\value{
Object of class \code{sampler.sergm}.
}
\description{
\code{sampler.sergm}  is used to specify MCMC samplers to sample signed networks 
from the distribution of Signed Exponential Random Graph Models. These samplers 
are needed to carry out MCMC estimation routines, quantify the uncertainty of the  
estimates and assess the goodness-of-fit of the estimated models.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: 
}}

