% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/sergm.R
\name{print.sergm}
\alias{print.sergm}
\alias{sergm}
\title{Exponential-Family Random Graph Models for Signed Networks}
\usage{
\method{print}{sergm}(x, ...)

sergm(formula, control = control.sergm())
}
\arguments{
\item{x}{See [print()].}

\item{...}{Other unused parameters
Automatically called when an object of class \code{\link{sergm}} is printed.}

\item{formula}{formula An \R \code{\link{formula}} object, of the form
\code{y ~ <model terms>}, where \code{y} is a numeric matrix including 
the signed entries 0, 1, and -1.}

\item{control}{list of parameters controling specifics of the MCMC}
}
\value{
Object of class \code{sergm}.
}
\description{
\code{\link{sergm}}  is used to fit signed exponential-family random graph models (SERGMs), 
in which the probability of a given network, y, on a set of nodes is 
exp\{theta s(y)\}/c(theta), s(y) is a vector of network statistics for y, 
theta is a natural parameter vector of the same length, 
and c(theta) is the normalizing constant for the distribution. 
\code{\link{sergm}} can return a maximum pseudo-likelihood estimate or 
an approximate maximum likelihood estimate based on a Monte Carlo scheme. 
There are currently four estimation algorithms implemented, which can be chosen 
in the \code{\link{control.sergm}} function that has to be supplied in the \code{control} parameter. 
\code{MLE} is the standard MCMCLE estimation routine 
with a log-normal approximation as proposed by Hummel et al. for binary networks. 
\code{Stepping} implements the stepping algorithm also introduced by by Hummel et al., 
while \code{RM} evokes the Robins Monroe algorithm for signed networks. 
Finally, the pseudo likelihood estimates based on a multinomial approximation 
of the intractable likelihood can be obtain through the option \code{MPLE}. 
For the pseudo likelihood the standard errors are approximated by a parametric 
bootstrap procedure introduced by Cranmer and Desmarais for static binary ERGMs.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: 
}}

