% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sergm.R
\name{summary.sergm}
\alias{summary.sergm}
\title{Summary Method for sergm Objects}
\usage{
\method{summary}{sergm}(object, ...)
}
\arguments{
\item{object}{A fitted sergm model (object of class \code{sergm}).}

\item{...}{Optional parameter (currently not in use).}
}
\value{
table of the results as a data.frame object.
}
\description{
Prints basic information of the results of an estimated sergm model.
}
