% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/tsergm.R
\name{print.summary.tsergm}
\alias{print.summary.tsergm}
\alias{summary.tsergm}
\title{Summary Method for tsergm Objects}
\usage{
\method{print}{summary.tsergm}(x, ...)

\method{summary}{tsergm}(object, ...)
}
\arguments{
\item{x}{See [print()].}

\item{...}{Optional parameter (currently not in use).}

\item{object}{A fitted sergm model (object of class \code{sergm}).}
}
\value{
table of the results as a data.frame object.
}
\description{
Prints basic information of the results of an estimated sergm model.
}
\section{Functions}{
\itemize{
\item \code{print.summary.tsergm}: 
}}

