% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsergm.R
\name{t_model_assessment}
\alias{t_model_assessment}
\title{Model assessment for tsergm Objects}
\usage{
t_model_assessment(
  object,
  sampler = ergm.sign::sampler.sergm(),
  cluster = NULL
)
}
\arguments{
\item{object}{A fitted sergm model (object of class \code{\link{sergm}}).}

\item{sampler}{A fitted sergm model (object of class \code{\link{sampler.sergm}}).}

\item{cluster}{A PSOCK or Fork cluster generated with \code{\link{makeCluster}} from the package \code{\link{parallel}}.}
}
\value{
numeric value of AIC.
}
\description{
Samples from the converged object and then calculate the distribution of 
edgewise shared friends and enemies as well as positive and negative degree distributions.
}
