% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/tsergm.R
\name{print.tsergm}
\alias{print.tsergm}
\alias{tsergm}
\title{Exponential-Family Random Graph Models for Dynamic Signed Networks}
\usage{
\method{print}{tsergm}(x, ...)

tsergm(formula, control = control.sergm())
}
\arguments{
\item{x}{See [print()].}

\item{...}{Other unused parameters
Automatically called when an object of class \code{\link{sergm}} is printed.}

\item{formula}{formula An \R \code{\link{formula}} object, of the form
\code{y ~ <model terms>}, where \code{y} is a numeric matrix including 
the signed entries 0, 1, and -1.}

\item{control}{list of parameters controlling specifics of the MCMC (see \code{\link{control.sergm}})}
}
\value{
Object of class \code{sergm}.
}
\description{
\code{\link{tsergm}}  is used to fit temporal signed exponential-family random graph models (TSERGMs), 
in which the probability of a given network at time point t, y_t, on a set of nodes is 
exp\{theta s(y_t, y_t-1)\}/c(theta, y_t-1), s(y_t, y_t-1) is a vector of network statistics for y_t, 
theta is a natural parameter vector of the same length, 
and c(theta) is the normalizing constant for the distribution. 
In the dynamic setting the sufficient statistics can also depend on earlier realizations of the networks 
and we are modelling the observed networks jointly. 
\code{\link{tsergm}} can return a maximum pseudo-likelihood estimate or 
an approximate maximum likelihood estimate based on a Monte Carlo scheme. 
There are currently four estimation algorithms implemented, which can be chosen 
in the \code{\link{control.sergm}} function that has to be supplied in the \code{control} parameter. 
\code{MLE} is the standard MCMCLE estimation routine 
with a log-normal approximation as proposed by Hummel et al. for binary networks. 
\code{Stepping} implements the stepping algorithm also introduced by by Hummel et al., 
while \code{RM} evokes the Robins Monroe algorithm for signed networks. 
Finally, the pseudo likelihood estimates based on a multinomial approximation 
of the intractable likelihood can be obtain through the option \code{MPLE}. 
For the pseudo likelihood the standard errors are approximated by a parametric 
bootstrap procedure introduced by Cranmer and Desmarais for static binary ERGMs.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: 
}}

